cool
Project with 3 people  in one week. M5Stack devices with a website using asynchronous web technology. Players can use the gyro sensors to control their movements in a maze game, with three balls appearing on each M5Stack for multiplayer gameplay. Additionally, the game state is synchronized in real-time on the website.